requirements.txt
pandas
numpy
seaborn
matplotlib
plotly
kagglehub
ipywidgets
sqlalchemy
psycopg2-binary
